+++ src/components/experts/index.ts
@@ -1,0 +1,4 @@
+export { ExpertCard } from "./ExpertCard";
+export { ExpertDetailsModal } from "./ExpertDetailsModal";
+export { AddEditExpertModal } from "./AddEditExpertModal";
+export { ExpertsSection } from "./ExpertsSection";